+++ src/Thingworx.d.ts
@@ -2095,13 +2095,13 @@ declare interface FieldDefinitionBase<key extends string, T = any> extends Field
 declare interface JSONInfoTable<T> {
     rows: T[];
     dataShape: {
-        fieldDefinitions: {[key in keyof T]: FieldDefinitionBase<key, T[key]>}
+        fieldDefinitions: {[key in keyof T]: key extends string ? FieldDefinitionBase<key, T[key]> : never}
     }
 }
 
 type NOTHING = void;
 type STRING = string;
-type NUMBER = string;
+type NUMBER = number;
 type BOOLEAN = boolean;
 
 type ANYSCALAR = unknown;
@@ -2177,7 +2177,7 @@ type THINGNAME<Template extends string | undefined = undefined, Shape extends st
  * A subclass of `TWRuntimeWidgets` that provides support for widget classes that can be defined using
  * the `BMCodeHost` extensions.
  */
- declare class TypescriptWidget extends TWRuntimeWidget {
+ declare abstract class TypescriptWidget extends TWRuntimeWidget {
     renderHtml(): string;
     afterRender(): void;